Implemented support for not failing reversing ADD INDEX
Also implemented support for not failing reversing ALTER TABLE CHANGE

+++ root/includes/mod_parser.php
@@ -82,7 +82,15 @@ class parser
 	*/
 	function reverse_query($orig_query)
 	{
-		if (preg_match('#ALTER TABLE\s([a-z_]+)\sADD(?:\sCOLUMN)?\s([a-z_]+)#i', $orig_query, $matches))
+		if (preg_match('#ALTER TABLE\s([a-z_]+)\sADD(?:\sCOLUMN)?\sINDEX\s\(([a-z_]+)\)#i', $orig_query, $matches))
+		{
+			return "SELECT NOW()";
+		}
+		else if (preg_match('#ALTER TABLE\s([a-z_]+)\sCHANGE#i', $orig_query, $matches))
+		{
+			return "SELECT NOW()";
+		}
+		else if (preg_match('#ALTER TABLE\s([a-z_]+)\sADD(?:\sCOLUMN)?\s([a-z_]+)#i', $orig_query, $matches))
 		{
 			return "ALTER TABLE {$matches[1]} DROP COLUMN {$matches[2]};";
 		}
